core/requirements.txt
django>=4.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
gspread>=5.12
google-auth>=2.23
google-auth-oauthlib>=1.1
